rag/requirements.txt
transformers
torch
sentence-transformers
scikit-learn
numpy
pandas
chromadb
langchain-community
langchain-core
langchain
valkey
python-dotenv
huggingface-hub
rich
